+++ NEWS.md
@@ -1,6 +1,11 @@
+
+# ggmagnify (development version)
+
+* New `corners` argument sets corner radius on target and inset. Thanks
+  to @thomasp85 for help with this.
+
 # ggmagnify 0.2.0
 
-* First CRAN release.
 * `from` and `to` can now be mapped to aesthetics, allowing different targets in
   different facets.
   - `from` can now be a logical vector of points to select.